+++ test/support/rj_runner.rb
@@ -1,7 +1,7 @@
 require 'open3'
 
-class RqRunner
-  RQ_BIN = ENV["RQ_BIN"] || File.expand_path("../../build/src/rq", __dir__)
+class RjRunner
+  RJ_BIN = ENV["RJ_BIN"] || File.expand_path("../../build/src/rj", __dir__)
 
   attr_reader :argv, :stdin, :stdout, :stderr, :status
 
@@ -11,7 +11,7 @@ class RqRunner
   end
 
   def run
-    @stdout, @stderr, process = Open3.capture3(*[RQ_BIN, *argv], stdin_data: stdin)
+    @stdout, @stderr, process = Open3.capture3(*[RJ_BIN, *argv], stdin_data: stdin)
     @status = process.exitstatus
   end
 end