+++ test/usage_from_readme.rb
@@ -1,28 +1,28 @@
-require_relative './support/rq_test'
+require_relative './support/rj_test'
 
-class UsageFromReadme < RqTest
-  rq_test! :first_element,
+class UsageFromReadme < RjTest
+  rj_test! :first_element,
       argv: [
         '.first',
         ],
         stdin:  :basic_array,
         stdout: :from_fixture
 
-  rq_test! :third_element,
+  rj_test! :third_element,
       argv: [
         '[2]',
         ],
         stdin:  :basic_array,
         stdout: :from_fixture
 
-  rq_test! :specific_fields,
+  rj_test! :specific_fields,
       argv: [
         '.first.slice("foo", "bar")',
         ],
         stdin:  :basic_array,
         stdout: :from_fixture
 
-  rq_test! :two_expressions,
+  rj_test! :two_expressions,
       argv: [
         '.first',
         '{ one: item["foo"], two: item.dig("bar", "name") }',
@@ -30,7 +30,7 @@ class UsageFromReadme < RqTest
         stdin:  :basic_array,
         stdout: :from_fixture
 
-  rq_test! :map,
+  rj_test! :map,
       argv: [
         '.map { |item| { one: item["foo"], two: item.dig("bar", "name") } }'
       ],